chore(management-express): add jsdoc
