+++ packages/management-express/src/types.ts
@@ -1,0 +1,158 @@
+import type { ManagementOptions } from '@monque/management';
+import type { Request, Response } from 'express';
+
+/**
+ * Express request objects passed to adapter callbacks.
+ *
+ * The adapter exposes both the incoming {@link Request} and outgoing {@link Response} so
+ * host applications can derive Management Surface context or OpenAPI server metadata from
+ * framework-native state such as headers, authenticated users, cookies, or mount details.
+ */
+export interface ManagementExpressContextInput {
+	/** Incoming Express request for the management route being handled. */
+	req: Request;
+	/** Outgoing Express response for the management route being handled. */
+	res: Response;
+}
+
+/**
+ * Creates request-specific context for the Management Surface from Express state.
+ *
+ * The returned value is passed through to `@monque/management` as `managementContext`.
+ * Management authorization hooks and payload serializers receive that same value as their
+ * `context` input. If omitted, the Management Surface receives no application context.
+ *
+ * @typeParam TContext - Application-specific context exposed to management hooks.
+ *
+ * @example Derive context from host authentication middleware.
+ * ```typescript
+ * app.use(
+ * 	'/monque',
+ * 	createManagementExpressRouter<{ userId: string; role: string }>({
+ * 		monque,
+ * 		context: ({ req }) => ({
+ * 			userId: req.get('x-user-id') ?? 'anonymous',
+ * 			role: req.get('x-role') ?? 'viewer',
+ * 		}),
+ * 		authorize: ({ action, context }) => {
+ * 			return context.role === 'operator' || action === 'read';
+ * 		},
+ * 	}),
+ * );
+ * ```
+ */
+export type ManagementExpressContextFactory<TContext = unknown> = (
+	input: ManagementExpressContextInput,
+) => TContext | Promise<TContext>;
+
+/**
+ * OpenAPI `servers[0].url` value or resolver for the mounted Express adapter.
+ *
+ * Use a string when the public URL is stable. Use a function when the URL depends on the
+ * current request, for example forwarded headers, tenant routing, or an Express mount path.
+ * Empty strings are normalized to `/`.
+ *
+ * @defaultValue The Express mount path, based on `req.baseUrl`.
+ *
+ * @example
+ * ```typescript
+ * createManagementExpressRouter({
+ * 	monque,
+ * 	openApi: {
+ * 		serverUrl: ({ req }) => {
+ * 			const host = req.get('x-forwarded-host') ?? req.get('host') ?? 'localhost';
+ * 			return `https://${host}${req.baseUrl}`;
+ * 		},
+ * 	},
+ * });
+ * ```
+ */
+export type ManagementExpressOpenApiServerUrl =
+	| string
+	| ((input: ManagementExpressContextInput) => string | Promise<string>);
+
+/**
+ * Configures adapter-served OpenAPI JSON.
+ *
+ * The OpenAPI document itself is generated by `@monque/management`; this Express adapter
+ * only chooses where to serve it and which mount-specific `servers` metadata to attach.
+ * Set {@link ManagementExpressRouterOptions.openApi} to `false` to disable this route.
+ */
+export interface ManagementExpressOpenApiOptions {
+	/**
+	 * Path where OpenAPI JSON is served, relative to the Express router mount.
+	 *
+	 * A leading slash is optional. For example, both `openapi.json` and `/openapi.json`
+	 * serve the document at the same router-relative path.
+	 *
+	 * @defaultValue `/openapi.json`
+	 */
+	path?: string;
+	/**
+	 * OpenAPI server URL for the mounted management API.
+	 *
+	 * This value is written to `document.servers` on the response. It does not change which
+	 * routes the adapter serves.
+	 *
+	 * @defaultValue The Express mount path, based on `req.baseUrl`.
+	 */
+	serverUrl?: ManagementExpressOpenApiServerUrl;
+}
+
+/**
+ * Options for creating an Express router backed by the Monque Management Surface.
+ *
+ * This extends the framework-neutral {@link ManagementOptions} from `@monque/management`
+ * with Express-specific request context creation and optional adapter-served OpenAPI JSON.
+ *
+ * The adapter does not implement authentication itself. Put host authentication middleware
+ * before the router, then use `context` and `authorize` for action-grained management
+ * authorization when needed.
+ *
+ * @typeParam TContext - Application-specific context exposed to management hooks.
+ *
+ * @example Mount the adapter under `/monque`.
+ * ```typescript
+ * import { createManagementExpressRouter } from '@monque/management-express';
+ * import express from 'express';
+ *
+ * const app = express();
+ *
+ * app.use(
+ * 	'/monque',
+ * 	requireOperator,
+ * 	createManagementExpressRouter({
+ * 		monque,
+ * 		context: ({ req }) => ({
+ * 			userId: req.get('x-user-id') ?? 'anonymous',
+ * 		}),
+ * 		authorize: ({ action, context }) => {
+ * 			return context.userId !== 'anonymous' || action === 'read';
+ * 		},
+ * 		openApi: {
+ * 			path: '/docs/openapi.json',
+ * 			serverUrl: 'https://ops.example.com/monque',
+ * 		},
+ * 	}),
+ * );
+ * ```
+ */
+export interface ManagementExpressRouterOptions<TContext = unknown>
+	extends ManagementOptions<TContext> {
+	/**
+	 * Optional factory that maps Express request state into Management Surface context.
+	 *
+	 * The returned value is available to management `authorize`, `serializePayload`, and
+	 * `serializePayloadByJobName` hooks as their `context` input.
+	 */
+	context?: ManagementExpressContextFactory<TContext>;
+	/**
+	 * Adapter-served OpenAPI JSON configuration.
+	 *
+	 * Pass `false` when the host application serves the generated OpenAPI document itself.
+	 * Pass an object to customize the router-relative path or OpenAPI server URL.
+	 *
+	 * @defaultValue `{ path: '/openapi.json', serverUrl: req.baseUrl }`
+	 */
+	openApi?: false | ManagementExpressOpenApiOptions;
+}
